refactor(parser): remove JS parser and update imports
