+++ src/services/__tests__/parseDoesNotThrow.browser.test.js
@@ -1,4 +1,4 @@
-import parseFundFile from '../parseFundFile';
+import parseFundFile from '@/utils/parseFundFile';
 import fs from 'fs';
 import path from 'path';
 import * as XLSX from 'xlsx';
