+++ src/services/dataLoader.js
@@ -92,9 +92,9 @@ export function ensureBenchmarkRows(list = []) {
         oneYear: null,
         threeYear: null,
         fiveYear: null,
-        sharpe: null,
-        stdDev5y: null,
-        expense: null
+        sharpe3Y: null,
+        stdDev5Y: null,
+        expenseRatio: null
       });
     } else {
       const row = map.get(key);
